Obtain EUI-64 from mbed HAL
The MCR20A does not have any built-in EUI-64, but this driver was
attempting to read the initial state of the address registers to get
one.

The address registers are initialised to ff:ff:ff:ff:ff:ff:ff:ff on
reset, but the read routine was broken - it read direct registers 7-14
when it should have read indirect registers 7-14. This meant it read a
slightly-random bitstring, as direct registers 12-14 are part of a
timer.

Modify the driver to get a 48-bit MAC address from `mbed_mac_address`
and convert that to a EUI-64 for use with radio.

Fixes #17.